refactor: 사용되지 않는 redux-toolkit 삭제 및 관리자 auth 검증을 쿠키를 사용하는 방식으로 변경
