+++ app/(admin)/admin/application/[id]/page.tsx
@@ -5,20 +5,18 @@ import { useEffect, useState } from 'react';
 import { getApplicationDetail } from '@/api';
 import { isAxiosError } from 'axios';
 import { ApplicationDetailType } from '@/types';
-import { APPLICATION_DETAIL_DEFAULT, ADMIN, MAIN_COLOR } from '@/constants';
+import { APPLICATION_DETAIL_DEFAULT, ADMIN, MAIN_COLOR, ACCESS_TOKEN } from '@/constants';
 import Loading from '@/components/Common/Loading';
-import { useAppDispatch } from '@/store';
 import { useRouter } from 'next/navigation';
-import { logout } from '@/store/adminSlice';
 import { useIsLoading } from '@/hooks/useIsLoading';
 import Application from '@/components/Admin/ApplicationDetail/Application';
+import { deleteCookie } from 'cookies-next';
 import * as S from './styled';
 
 const ApplicationDetail = ({ params: { id } }: { params: { id: number } }) => {
   const isLoading = useIsLoading();
   const [application, setApplication] = useState<ApplicationDetailType>(APPLICATION_DETAIL_DEFAULT);
   const router = useRouter();
-  const dispatch = useAppDispatch();
   const color = MAIN_COLOR;
 
   useEffect(() => {
@@ -32,7 +30,8 @@ const ApplicationDetail = ({ params: { id } }: { params: { id: number } }) => {
   }, [id, isLoading]);
 
   const logoutHandler = () => {
-    dispatch(logout());
+    // dispatch(logout());
+    deleteCookie(ACCESS_TOKEN);
     router.push(ADMIN.LOGIN);
   };
