+++ app/(user)/styled/index.ts
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+export const PointerWrapper = styled.main`
+  width: 100%;
+  height: auto;
+
+  cursor: none;
+`;
