+++ app/lib/Provider/index.tsx
@@ -1,3 +1,2 @@
-export { default as RTProvider } from './RTProvider';
 export { default as StyledProvider } from './StyledProvider';
 export { default as NextAuthProvider } from './SessionProvider';
