+++ constants/login.ts
@@ -1,4 +1,4 @@
-export const LOGIN = {
+export const LOGIN_DEFAULT_VALUE = {
   id: '',
   password: '',
 } as const;
