+++ package.json
@@ -3,7 +3,6 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@reduxjs/toolkit": "^1.9.5",
     "@tanstack/react-query": "^5.17.19",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
@@ -20,8 +19,6 @@
     "react": "^18.2.0",
     "react-device-detect": "^2.2.3",
     "react-dom": "^18.2.0",
-    "react-redux": "^8.0.5",
-    "redux-persist": "^6.0.0",
     "styled-components": "^6.0.7",
     "sweetalert2": "^11.7.22",
     "web-vitals": "^2.1.4"
@@ -54,7 +51,6 @@
     "@next/bundle-analyzer": "^14.1.0",
     "@svgr/webpack": "^8.0.1",
     "@types/facepaint": "^1.2.2",
-    "@types/react-redux": "^7.1.33",
     "@typescript-eslint/parser": "^6.19.1",
     "compression-webpack-plugin": "^11.0.0",
     "eslint": "^8.33.0",
