--- utils/LocalStorage/index.ts
@@ -1,20 +1,0 @@
-export class LocalStorage {
-  static setItem(key: string, value: string) {
-    if (typeof window !== 'undefined') {
-      localStorage.setItem(key, value);
-    }
-  }
-
-  static getItem(key: string) {
-    if (typeof window !== 'undefined') {
-      return localStorage.getItem(key);
-    }
-    return null;
-  }
-
-  static removeItem(key: string) {
-    if (typeof window !== 'undefined') {
-      localStorage.removeItem(key);
-    }
-  }
-}
