+++ utils/index.tsx
@@ -2,5 +2,4 @@ export * from './Validator';
 export * from './DeviceChecker';
 export * from './Search';
 export * from './Formatter';
-export * from './LocalStorage';
 export * from './Alert';
